# be/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Thăng Long Chè Việt — Storefront API",
        "description": "Public REST API for Nuxt storefront: products, blog, cart, orders, booking, contact. Admin UI uses Filament at /admin.",
        "contact": {
            "email": "admin@thanglongcheviet.vn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8002",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/customer/appointments": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "List my appointments",
                "operationId": "customerAppointments",
                "responses": {
                    "200": {
                        "description": "Appointment list"
                    }
                },
                "security": [
                    {
                        "customerBearer": []
                    }
                ]
            }
        },
        "/api/customer/appointments/{id}": {
            "delete": {
                "tags": [
                    "Customer"
                ],
                "summary": "Cancel appointment (>24h before scheduled time)",
                "operationId": "cancelCustomerAppointment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cancelled"
                    },
                    "400": {
                        "description": "Cannot cancel"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "customerBearer": []
                    }
                ]
            }
        },
        "/api/customer/register": {
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Register customer account",
                "operationId": "customerRegister",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "phone",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 6
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registered"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/customer/login": {
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Customer login",
                "operationId": "customerLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "password"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authenticated"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/customer/logout": {
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Revoke customer token",
                "operationId": "customerLogout",
                "responses": {
                    "200": {
                        "description": "Logged out"
                    }
                },
                "security": [
                    {
                        "customerBearer": []
                    }
                ]
            }
        },
        "/api/customer/profile": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Customer profile",
                "operationId": "customerProfile",
                "responses": {
                    "200": {
                        "description": "Profile"
                    }
                },
                "security": [
                    {
                        "customerBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "customerBearer": {
                "type": "http",
                "description": "Customer API token (not admin token)",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}